test: adjust tests

+++ cellseg_models_pytorch/transforms/tests/test_norm.py
@@ -13,7 +13,7 @@ from cellseg_models_pytorch.transforms.functional.normalization import (
 @pytest.mark.parametrize("lower", [0.0, 0.1])
 @pytest.mark.parametrize("upper", [99.99, 100.0])
 def test_percentile_normalize(img_sample, lower, upper) -> None:
-    im = img_sample
+    im = img_sample.astype(np.float32)
     nim = percentile_normalize(im, lower, upper)
 
     assert nim.dtype == "float32"
@@ -24,7 +24,7 @@ def test_percentile_normalize(img_sample, lower, upper) -> None:
 @pytest.mark.parametrize("amin", [None, 0.0])
 @pytest.mark.parametrize("amax", [1.0, None])
 def test_percentile_normalize99(img_sample, amin, amax) -> None:
-    im = img_sample
+    im = img_sample.astype(np.float32)
     nim = percentile_normalize99(im, amin, amax)
 
     assert nim.dtype == "float32"
@@ -35,7 +35,7 @@ def test_percentile_normalize99(img_sample, amin, amax) -> None:
 @pytest.mark.parametrize("amax", [1.0, None])
 @pytest.mark.parametrize("standardize", [True, False])
 def test_normalize(img_sample, standardize, amin, amax) -> None:
-    im = img_sample
+    im = img_sample.astype(np.float32)
     nim = normalize(im, standardize, amin, amax)
 
     assert nim.dtype == "float32"
@@ -45,7 +45,7 @@ def test_normalize(img_sample, standardize, amin, amax) -> None:
 @pytest.mark.parametrize("amin", [None, 0.0])
 @pytest.mark.parametrize("amax", [1.0, None])
 def test_minmax_normalize(img_sample, amin, amax) -> None:
-    im = img_sample
+    im = img_sample.astype(np.float32)
     nim = minmax_normalize(im, amin, amax)
 
     assert nim.dtype == "float32"
@@ -55,7 +55,7 @@ def test_minmax_normalize(img_sample, amin, amax) -> None:
 
 @pytest.mark.parametrize("norm", [True, False])
 def test_float2ubyte(img_sample, norm) -> None:
-    im = img_sample
+    im = img_sample.astype(np.float32)
     nim = normalize(im)
     ubyte = float2ubyte(nim, norm)
